+++ CurrencyExchange/app/src/main/java/com/application/currency/exchange/domain/entity/model/ConversionValue.kt
@@ -1,4 +1,4 @@
 package com.application.currency.exchange.domain.entity.model
 
-data class ConversionValue(val value: Float,
+data class ConversionValue(val value: Double,
                            val date: String)
