+++ CurrencyExchange/app/src/main/java/com/application/currency/exchange/domain/entity/model/ExchangeRateInfo.kt
@@ -1,5 +1,5 @@
 package com.application.currency.exchange.domain.entity.model
 
 data class ExchangeRateInfo(val rates: List<Rate>, val sellRate: Rate? = null,
-                            val receiveRate: Rate? = null, val sellValue: Float? = null,
-                            val receiveValue: Float? = null)
+                            val receiveRate: Rate? = null, val sellValue: Double? = null,
+                            val receiveValue: Double? = null)
