+++ CurrencyExchange/app/src/main/java/com/application/currency/exchange/domain/entity/model/Rate.kt
@@ -8,5 +8,5 @@ const val TABLE_NAME = "table_rate"
 @Entity(tableName = TABLE_NAME)
 data class Rate(val currency: String,
                 val conversionMap: MutableMap<String, ConversionValue>,
-                var amount: Float = 0f,
+                var amount: Double = 0.0,
                 @PrimaryKey(autoGenerate = true) val id: Int = 0)
