+++ CurrencyExchange/app/src/main/java/com/application/currency/exchange/presentation/event/MainScreenEvent.kt
@@ -4,7 +4,7 @@ import com.application.currency.exchange.domain.entity.model.Rate
 
 sealed class MainScreenEvent: BaseScreenEvent() {
     data object OnGetExchangeRate: MainScreenEvent()
-    data class OnUpdateSellValue(val amount: Float): MainScreenEvent()
+    data class OnUpdateSellValue(val amount: Double): MainScreenEvent()
     data class OnUpdateSellCurrency(val rate: Rate): MainScreenEvent()
     data class OnUpdateReceiveCurrency(val rate: Rate): MainScreenEvent()
     data object OnUpdateBalance: MainScreenEvent()
